+++ src/lib/pages/interact-contract/types.ts
@@ -1,0 +1,19 @@
+import { z } from "zod";
+
+import { ContractInteractionTabs, zBechAddr32 } from "lib/types";
+
+export const zInteractContractQueryParams = z.object({
+  selectedType: z.union([
+    z.nativeEnum(ContractInteractionTabs),
+    z
+      .string()
+      .optional()
+      .transform(() => ContractInteractionTabs.Query),
+  ]),
+  contract: zBechAddr32.default(""),
+  msg: z.string().default(""),
+});
+
+export type InteractContractQueryParams = z.infer<
+  typeof zInteractContractQueryParams
+>;